+++ server/src/Hiscary.Stories.IntegrationEvents/Outgoing/StoryInformationRefreshChunkProcessedIntegrationEvent.cs
@@ -6,11 +6,11 @@ namespace Hiscary.Stories.IntegrationEvents.Outgoing;
 public sealed class StoryInformationRefreshChunkProcessedIntegrationEvent(
     Guid RequesterId,
     StoryChunk[] Stories,
-    long Index,
-    long ChunkSize) : BaseIntegrationEvent
+    int Index,
+    int ChunkSize) : BaseIntegrationEvent
 {
     public Guid RequesterId { get; } = RequesterId;
     public StoryChunk[] Stories { get; } = Stories;
-    public long Index { get; } = Index;
-    public long ChunkSize { get; } = ChunkSize;
+    public int Index { get; } = Index;
+    public int ChunkSize { get; } = ChunkSize;
 }
